add `ui/volume-slider` tests
